- partial work on being able to edit and view params on nodes - adding in a hint system - the hint system will initially just handle interpreting ints as some enum type, allowing to gui to display a dropdown with text instead of a number editor

+++ src/components/computes/outputcompute.cpp
@@ -54,7 +54,7 @@ void OutputCompute::save(SimpleSaver& saver) const {
   Compute::save(saver);
 
   // Serialize the param type.
-  size_t type = _param_type;
+  size_t type = static_cast<size_t>(_param_type);
   saver.save(type);
 
   // Serialize the exposed value.
